+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Add subexamples support in HTML and TeX.
 
+- Remove automatic quotes in translation to allow for custom comments.
+
 # v0.0.1
 
 - Develop interlinear extension.