+++ NEWS.md
@@ -1,10 +1,12 @@
-# v0.0.1-9000
+# v0.0.2
 
 - Add numbering in HTML. 
 
 - Add subexamples support in HTML and TeX.
 
 - Remove automatic quotes in translation to allow for custom comments.
+
+- Add support for custom abbreviations in HTML output.
 
 # v0.0.1